SampoGallery: Improvements to title
Changed SAMPO to a conditional title: If the entry has no article-id
in kansallisbiografia, the title is LINKED DATA FINLAND (where the
text is a link to the event entry from ldf.fi/history). If the entry
has an article-id, the title is:
KANSALLISBIOGRAFIA via LINKED DATA FINLAND
where the first part is a link to the entry in kansallisbiografia
and the second part is a link to ldf.fi/history as above.
Also some editorial cleanup.