fix: Further improvements to year range in EventList
Instead of min(time)-max(time) as in the previous version, this
uses GROUP_CONCAT with DISTINCT. The previous solution was showing
also single years as ranges (1911-1911), this one doesn't force a
single year to a range. Works as long as all the descriptions of a
statement refer to the same range of years.